+++ electron/Index.vue
@@ -293,8 +293,15 @@
             });
 
             window.addEventListener('keydown', (e) => {
-                if(getKey(e) === Keys.Tab && e.ctrlKey && !e.altKey && !e.shiftKey)
-                    parent.send('switch-tab', this.character);
+                const key = getKey(e);
+
+                if ((key === Keys.Tab) && (e.ctrlKey) && (!e.altKey)) {
+                    parent.send(`${e.shiftKey ? 'previous' : 'switch'}-tab`, this.character);
+                }
+
+                if (((key === Keys.PageDown) || (key === Keys.PageUp)) && (e.ctrlKey) && (!e.altKey) && (!e.shiftKey)) {
+                  parent.send(`${key === Keys.PageUp ? 'previous' : 'switch'}-tab`, this.character);
+                }
             });
 
             log.debug('init.chat.listeners.done');